Upgrade our CI/CD process by improving docker build process and switching to circleci 2.0
