+++ Makefile
@@ -10,7 +10,7 @@ SHELL := /bin/bash
 REPO := $(shell pwd)
 GOFILES_NOVENDOR := $(shell find ${REPO} -type f -name '*.go' -not -path "${REPO}/vendor/*")
 PACKAGES_NOVENDOR := $(shell go list github.com/hyperledger/burrow/... | grep -v /vendor/)
-VERSION := $(shell cat ${REPO}/version/version.go | tail -n 1 | cut -d \  -f 4 | tr -d '"')
+VERSION := $(shell go run ./util/version/cmd/main.go)
 VERSION_MIN := $(shell echo ${VERSION} | cut -d . -f 1-2)
 COMMIT_SHA := $(shell echo `git rev-parse --short --verify HEAD`)
 
@@ -20,6 +20,10 @@ DOCKER_NAMESPACE := quay.io/monax
 .PHONY: greet
 greet:
 	@echo "Hi! I'm the marmot that will help you with burrow v${VERSION}"
+
+.PHONY: version
+version:
+	@echo "${VERSION}"
 
 ### Formatting, linting and vetting
 
@@ -117,8 +121,12 @@ build_race_client:
 
 # test burrow
 .PHONY: test
-test: build
-	@go test ${PACKAGES_NOVENDOR} -tags integration
+test:
+	@go test ${PACKAGES_NOVENDOR}
+
+.PHONY: test_integration
+test_integration:
+	@go test ./rpc/tendermint/test -tags integration
 
 # test burrow with checks for race conditions
 .PHONY: test_race
@@ -130,23 +138,7 @@ test_race: build_race
 # build docker image for burrow
 .PHONY: build_docker_db
 build_docker_db: check
-	@mkdir -p ${REPO}/target/docker
-	docker build -t ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA} ${REPO}
-	docker run --rm --entrypoint cat ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA} /usr/local/bin/burrow > ${REPO}/target/docker/burrow.dockerartefact
-	docker run --rm --entrypoint cat ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA} /usr/local/bin/burrow-client > ${REPO}/target/docker/burrow-client.dockerartefact
-	docker build -t ${DOCKER_NAMESPACE}/db:${VERSION} -f Dockerfile.deploy ${REPO}
-
-	@rm ${REPO}/target/docker/burrow.dockerartefact
-	@rm ${REPO}/target/docker/burrow-client.dockerartefact
-	docker rmi ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA}
-
-### Test docker images for github.com/hyperledger/burrow
-
-# test docker image for burrow
-.PHONY: test_docker_db
-test_docker_db: check
-	docker build -t ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA} ${REPO}
-	docker run ${DOCKER_NAMESPACE}/db:build-${COMMIT_SHA} glide nv | xargs go test -tags integration
+	@./build_tool.sh
 
 ### Clean up
